Move header tips to the header
